#185 Koitettu korjata testit
